+++ client/ssr/Asset.tsx
