+++ examples/ssr/src/pages/__root.tsx
@@ -1,5 +1,5 @@
 import { Link, Outlet, createRootRoute } from '@umijs/tnf/router';
-import { Meta } from '@umijs/tnf/ssr';
+import { Meta } from '../../../../client/ssr/ssr';
 
 export const Route = createRootRoute({
   component: RootComponent,
